Refactor Mengganti nama file models dan controller menjadi snakecase dan Mempersiapkan file untuk merge dengan frontend pada main
