+++ lib/Models/Avatar/avatar_detail.dart
