+++ lib/controllers/inventory_controller.dart
@@ -3,38 +3,39 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:http/http.dart' as http;
-import 'package:parkquest_parkir_gamifikasi/Controllers/ProfileController.dart';
-import 'package:parkquest_parkir_gamifikasi/Models/ParkSearch/ParkArea.dart';
-import 'package:parkquest_parkir_gamifikasi/Models/ParkSearch/ParkData.dart';
-import 'package:parkquest_parkir_gamifikasi/Models/ParkSearch/ParkRecommendation.dart';
+import 'package:parkquest_parkir_gamifikasi/controllers/leaderboard_controller.dart';
+import 'package:parkquest_parkir_gamifikasi/controllers/profile_controller.dart';
+import 'package:parkquest_parkir_gamifikasi/models/avatar/user_avatar.dart';
 import 'package:parkquest_parkir_gamifikasi/constants.dart';
 
-class ParkSearchController extends GetxController {
-  Rx<List> datasParkArea = Rx<List>([]);
-  Rx<List> datasParkData = Rx<List>([]);
-  Rx<List> datasParkRecommendation = Rx<List>([]);
+class InventoryController extends GetxController {
+  Rx<List<UserAvatarModel>> datasBasic = Rx<List<UserAvatarModel>>([]);
+  Rx<List<UserAvatarModel>> datasRare = Rx<List<UserAvatarModel>>([]);
+  Rx<List<UserAvatarModel>> datasLegendary = Rx<List<UserAvatarModel>>([]);
   final isLoading = false.obs;
   final box = GetStorage();
-  Rxn<ParkAreaModel> parkAreaData = Rxn<ParkAreaModel>();
-  Rxn<ParkRecommendationModel> parkRecommendationData =
-      Rxn<ParkRecommendationModel>();
 
   // User
   final ProfileController _profilecontroller = Get.put(ProfileController());
+  // User Leaderboard
+  final LeaderboardController _leaderboardcontroller =
+      Get.put(LeaderboardController());
 
   @override
   void onInit() {
     super.onInit();
-    parkArea();
+    inventoryBasic();
+    inventoryRare();
+    inventoryLegendary();
   }
 
-  Future parkArea() async {
+  Future inventoryBasic() async {
     try {
       isLoading.value = true;
       final token = box.read('token');
 
       final response = await http.get(
-        Uri.parse('${apiUrl}parkAreaSearch'),
+        Uri.parse('${apiUrl}inventory/basic'),
         headers: {
           ...headers,
           'Authorization': 'Bearer $token',
@@ -44,11 +45,11 @@ class ParkSearchController extends GetxController {
       final content = json.decode(response.body);
 
       if (response.statusCode == 200) {
-        datasParkArea.value.clear();
+        datasBasic.value.clear();
         isLoading.value = false;
 
         for (var item in content['data']) {
-          datasParkArea.value.add(ParkAreaModel.fromJson(item));
+          datasBasic.value.add(UserAvatarModel.fromJson(item));
         }
 
         debugPrint(content.toString());
@@ -72,13 +73,13 @@ class ParkSearchController extends GetxController {
     }
   }
 
-  Future parkData({required String parkAreaId}) async {
+  Future inventoryRare() async {
     try {
       isLoading.value = true;
       final token = box.read('token');
 
       final response = await http.get(
-        Uri.parse('${apiUrl}parkData/$parkAreaId'),
+        Uri.parse('${apiUrl}inventory/rare'),
         headers: {
           ...headers,
           'Authorization': 'Bearer $token',
@@ -88,16 +89,11 @@ class ParkSearchController extends GetxController {
       final content = json.decode(response.body);
 
       if (response.statusCode == 200) {
-        datasParkData.value.clear();
+        datasRare.value.clear();
         isLoading.value = false;
 
-        final dataArea = content['dataParkArea'];
-        parkAreaData.value = ParkAreaModel.fromJson(dataArea);
-
-        Get.offAllNamed('/detailPencarianParkir');
-
         for (var item in content['data']) {
-          datasParkData.value.add(ParkDataModel.fromJson(item));
+          datasRare.value.add(UserAvatarModel.fromJson(item));
         }
 
         debugPrint(content.toString());
@@ -121,13 +117,13 @@ class ParkSearchController extends GetxController {
     }
   }
 
-  Future parkRecommendation({required String parkAreaId}) async {
+  Future inventoryLegendary() async {
     try {
       isLoading.value = true;
       final token = box.read('token');
 
       final response = await http.get(
-        Uri.parse('${apiUrl}parkRecommendation/$parkAreaId'),
+        Uri.parse('${apiUrl}inventory/legendary'),
         headers: {
           ...headers,
           'Authorization': 'Bearer $token',
@@ -137,14 +133,11 @@ class ParkSearchController extends GetxController {
       final content = json.decode(response.body);
 
       if (response.statusCode == 200) {
-        datasParkRecommendation.value.clear();
+        datasLegendary.value.clear();
         isLoading.value = false;
 
-        Get.offAllNamed('/daftarRekomendasiParkir');
-
         for (var item in content['data']) {
-          datasParkRecommendation.value
-              .add(ParkRecommendationModel.fromJson(item));
+          datasLegendary.value.add(UserAvatarModel.fromJson(item));
         }
 
         debugPrint(content.toString());
@@ -168,61 +161,13 @@ class ParkSearchController extends GetxController {
     }
   }
 
-  Future parkRecommendationDetail(
-      {required String parkRecommendationId}) async {
-    try {
-      isLoading.value = true;
-      final token = box.read('token');
-
-      final response = await http.get(
-        Uri.parse('${apiUrl}parkRecommendationDetail/$parkRecommendationId'),
-        headers: {
-          ...headers,
-          'Authorization': 'Bearer $token',
-        },
-      );
-
-      final content = json.decode(response.body);
-
-      if (response.statusCode == 200) {
-        parkRecommendationData.value = null;
-        isLoading.value = false;
-
-        final dataRecommendation = content['data'];
-        parkRecommendationData.value =
-            ParkRecommendationModel.fromJson(dataRecommendation);
-
-        Get.offAllNamed('/detailDaftarRekomendasiParkir');
-
-        debugPrint(content.toString());
-      } else {
-        isLoading.value = false;
-
-        Get.snackbar(
-          'Error',
-          content['message'],
-          snackPosition: SnackPosition.BOTTOM,
-          backgroundColor: Colors.red,
-          colorText: Colors.white,
-        );
-
-        debugPrint(content.toString());
-      }
-    } catch (e) {
-      isLoading.value = false;
-
-      debugPrint(e.toString());
-    }
-  }
-
-  Future parkRecommendationAccepted(
-      {required String parkRecommendationId}) async {
+  Future updateAvatar({required String avatarId}) async {
     try {
       isLoading.value = true;
       final token = box.read('token');
 
       final response = await http.post(
-        Uri.parse('${apiUrl}parkRecommendationAccepted/$parkRecommendationId'),
+        Uri.parse('${apiUrl}inventory/updateAvatar/$avatarId'),
         headers: {
           ...headers,
           'Authorization': 'Bearer $token',
@@ -235,6 +180,7 @@ class ParkSearchController extends GetxController {
         isLoading.value = false;
 
         await _profilecontroller.profile();
+        await _leaderboardcontroller.userLeaderboard();
 
         Get.offAllNamed('/dashboard');
