+++ lib/controllers/leaderboard_controller.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:http/http.dart' as http;
-import 'package:parkquest_parkir_gamifikasi/Models/Leaderboard/Leaderboard.dart';
-import 'package:parkquest_parkir_gamifikasi/Models/Leaderboard/UserLeaderboard.dart';
+import 'package:parkquest_parkir_gamifikasi/models/leaderboard/leaderboard.dart';
+import 'package:parkquest_parkir_gamifikasi/models/leaderboard/user_leaderboard.dart';
 import 'package:parkquest_parkir_gamifikasi/constants.dart';
 
 class LeaderboardController extends GetxController {
