+++ lib/controllers/park_recommendation_controller.dart
@@ -6,8 +6,8 @@ import 'package:get_storage/get_storage.dart';
 import 'package:http/http.dart' as http;
 import 'package:image_picker/image_picker.dart';
 import 'package:parkquest_parkir_gamifikasi/constants.dart';
-import 'package:parkquest_parkir_gamifikasi/Models/ParkRecommendation/ParkArea.dart';
-import 'package:parkquest_parkir_gamifikasi/Controllers/ProfileController.dart';
+import 'package:parkquest_parkir_gamifikasi/models/park_recommendation/park_area.dart';
+import 'package:parkquest_parkir_gamifikasi/controllers/profile_controller.dart';
 
 class ParkRecommendationController extends GetxController {
   Rx<List> datas = Rx<List>([]);
