+++ lib/controllers/park_search_controller.dart
@@ -3,39 +3,38 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:http/http.dart' as http;
-import 'package:parkquest_parkir_gamifikasi/Controllers/LeaderboardController.dart';
-import 'package:parkquest_parkir_gamifikasi/Controllers/ProfileController.dart';
-import 'package:parkquest_parkir_gamifikasi/Models/Avatar/UserAvatar.dart';
+import 'package:parkquest_parkir_gamifikasi/controllers/profile_controller.dart';
+import 'package:parkquest_parkir_gamifikasi/models/park_search/park_area.dart';
+import 'package:parkquest_parkir_gamifikasi/models/park_search/park_data.dart';
+import 'package:parkquest_parkir_gamifikasi/models/park_search/park_recommendation.dart';
 import 'package:parkquest_parkir_gamifikasi/constants.dart';
 
-class InventoryController extends GetxController {
-  Rx<List<UserAvatarModel>> datasBasic = Rx<List<UserAvatarModel>>([]);
-  Rx<List<UserAvatarModel>> datasRare = Rx<List<UserAvatarModel>>([]);
-  Rx<List<UserAvatarModel>> datasLegendary = Rx<List<UserAvatarModel>>([]);
+class ParkSearchController extends GetxController {
+  Rx<List> datasParkArea = Rx<List>([]);
+  Rx<List> datasParkData = Rx<List>([]);
+  Rx<List> datasParkRecommendation = Rx<List>([]);
   final isLoading = false.obs;
   final box = GetStorage();
+  Rxn<ParkAreaModel> parkAreaData = Rxn<ParkAreaModel>();
+  Rxn<ParkRecommendationModel> parkRecommendationData =
+      Rxn<ParkRecommendationModel>();
 
   // User
   final ProfileController _profilecontroller = Get.put(ProfileController());
-  // User Leaderboard
-  final LeaderboardController _leaderboardcontroller =
-      Get.put(LeaderboardController());
 
   @override
   void onInit() {
     super.onInit();
-    inventoryBasic();
-    inventoryRare();
-    inventoryLegendary();
+    parkArea();
   }
 
-  Future inventoryBasic() async {
+  Future parkArea() async {
     try {
       isLoading.value = true;
       final token = box.read('token');
 
       final response = await http.get(
-        Uri.parse('${apiUrl}inventory/basic'),
+        Uri.parse('${apiUrl}parkAreaSearch'),
         headers: {
           ...headers,
           'Authorization': 'Bearer $token',
@@ -45,11 +44,11 @@ class InventoryController extends GetxController {
       final content = json.decode(response.body);
 
       if (response.statusCode == 200) {
-        datasBasic.value.clear();
+        datasParkArea.value.clear();
         isLoading.value = false;
 
         for (var item in content['data']) {
-          datasBasic.value.add(UserAvatarModel.fromJson(item));
+          datasParkArea.value.add(ParkAreaModel.fromJson(item));
         }
 
         debugPrint(content.toString());
@@ -73,13 +72,13 @@ class InventoryController extends GetxController {
     }
   }
 
-  Future inventoryRare() async {
+  Future parkData({required String parkAreaId}) async {
     try {
       isLoading.value = true;
       final token = box.read('token');
 
       final response = await http.get(
-        Uri.parse('${apiUrl}inventory/rare'),
+        Uri.parse('${apiUrl}parkData/$parkAreaId'),
         headers: {
           ...headers,
           'Authorization': 'Bearer $token',
@@ -89,11 +88,16 @@ class InventoryController extends GetxController {
       final content = json.decode(response.body);
 
       if (response.statusCode == 200) {
-        datasRare.value.clear();
+        datasParkData.value.clear();
         isLoading.value = false;
 
+        final dataArea = content['dataParkArea'];
+        parkAreaData.value = ParkAreaModel.fromJson(dataArea);
+
+        Get.offAllNamed('/detailPencarianParkir');
+
         for (var item in content['data']) {
-          datasRare.value.add(UserAvatarModel.fromJson(item));
+          datasParkData.value.add(ParkDataModel.fromJson(item));
         }
 
         debugPrint(content.toString());
@@ -117,13 +121,13 @@ class InventoryController extends GetxController {
     }
   }
 
-  Future inventoryLegendary() async {
+  Future parkRecommendation({required String parkAreaId}) async {
     try {
       isLoading.value = true;
       final token = box.read('token');
 
       final response = await http.get(
-        Uri.parse('${apiUrl}inventory/legendary'),
+        Uri.parse('${apiUrl}parkRecommendation/$parkAreaId'),
         headers: {
           ...headers,
           'Authorization': 'Bearer $token',
@@ -133,11 +137,14 @@ class InventoryController extends GetxController {
       final content = json.decode(response.body);
 
       if (response.statusCode == 200) {
-        datasLegendary.value.clear();
+        datasParkRecommendation.value.clear();
         isLoading.value = false;
 
+        Get.offAllNamed('/daftarRekomendasiParkir');
+
         for (var item in content['data']) {
-          datasLegendary.value.add(UserAvatarModel.fromJson(item));
+          datasParkRecommendation.value
+              .add(ParkRecommendationModel.fromJson(item));
         }
 
         debugPrint(content.toString());
@@ -161,13 +168,61 @@ class InventoryController extends GetxController {
     }
   }
 
-  Future updateAvatar({required String avatarId}) async {
+  Future parkRecommendationDetail(
+      {required String parkRecommendationId}) async {
+    try {
+      isLoading.value = true;
+      final token = box.read('token');
+
+      final response = await http.get(
+        Uri.parse('${apiUrl}parkRecommendationDetail/$parkRecommendationId'),
+        headers: {
+          ...headers,
+          'Authorization': 'Bearer $token',
+        },
+      );
+
+      final content = json.decode(response.body);
+
+      if (response.statusCode == 200) {
+        parkRecommendationData.value = null;
+        isLoading.value = false;
+
+        final dataRecommendation = content['data'];
+        parkRecommendationData.value =
+            ParkRecommendationModel.fromJson(dataRecommendation);
+
+        Get.offAllNamed('/detailDaftarRekomendasiParkir');
+
+        debugPrint(content.toString());
+      } else {
+        isLoading.value = false;
+
+        Get.snackbar(
+          'Error',
+          content['message'],
+          snackPosition: SnackPosition.BOTTOM,
+          backgroundColor: Colors.red,
+          colorText: Colors.white,
+        );
+
+        debugPrint(content.toString());
+      }
+    } catch (e) {
+      isLoading.value = false;
+
+      debugPrint(e.toString());
+    }
+  }
+
+  Future parkRecommendationAccepted(
+      {required String parkRecommendationId}) async {
     try {
       isLoading.value = true;
       final token = box.read('token');
 
       final response = await http.post(
-        Uri.parse('${apiUrl}inventory/updateAvatar/$avatarId'),
+        Uri.parse('${apiUrl}parkRecommendationAccepted/$parkRecommendationId'),
         headers: {
           ...headers,
           'Authorization': 'Bearer $token',
@@ -180,7 +235,6 @@ class InventoryController extends GetxController {
         isLoading.value = false;
 
         await _profilecontroller.profile();
-        await _leaderboardcontroller.userLeaderboard();
 
         Get.offAllNamed('/dashboard');
