+++ lib/views/kode_referral/bagikan.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 // import 'package:get_storage/get_storage.dart';
 import 'package:get/get.dart';
-import 'package:parkquest_parkir_gamifikasi/Controllers/ReferralCodeController.dart';
+import 'package:parkquest_parkir_gamifikasi/controllers/referral_code_controller.dart';
 
 class KodeReferralBagikan extends StatelessWidget {
   KodeReferralBagikan({super.key});
