+++ lib/views/pencarian_parkir/daftar_rekomendasi_parkir.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 // import 'package:get_storage/get_storage.dart';
-import 'package:parkquest_parkir_gamifikasi/Controllers/ParkSearchController.dart';
-import 'package:parkquest_parkir_gamifikasi/Models/ParkSearch/ParkRecommendation.dart';
+import 'package:parkquest_parkir_gamifikasi/controllers/park_search_controller.dart';
+import 'package:parkquest_parkir_gamifikasi/models/park_search/park_recommendation.dart';
 import 'package:parkquest_parkir_gamifikasi/constants.dart';
 
 class DaftarRekomendasiParkir extends StatelessWidget {
