+++ lib/views/profile/setting.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 // import 'package:get_storage/get_storage.dart';
 import 'package:get/get.dart';
-import 'package:parkquest_parkir_gamifikasi/Controllers/AuthenticationController.dart';
+import 'package:parkquest_parkir_gamifikasi/controllers/authentication_controller.dart';
 
 class Setting extends StatelessWidget {
   Setting({super.key});
