+++ lib/views/shop/detail_shop.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 // import 'package:get_storage/get_storage.dart';
 import 'package:get/get.dart';
-import 'package:parkquest_parkir_gamifikasi/Controllers/ShopController.dart';
+import 'package:parkquest_parkir_gamifikasi/controllers/shop_controller.dart';
 import 'package:parkquest_parkir_gamifikasi/constants.dart';
 
 class DetailShop extends StatelessWidget {
