+++ lib/views/survey/detail_survey.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 // import 'package:get_storage/get_storage.dart';
 import 'package:get/get.dart';
-import 'package:parkquest_parkir_gamifikasi/Controllers/SurveyController.dart';
-import 'package:parkquest_parkir_gamifikasi/Models/Survey/SurveyDetail.dart';
+import 'package:parkquest_parkir_gamifikasi/controllers/survey_controller.dart';
+import 'package:parkquest_parkir_gamifikasi/models/survey/survey_detail.dart';
 import 'package:parkquest_parkir_gamifikasi/constants.dart';
 
 class DetailSurvey extends StatelessWidget {
